requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
reportlab==4.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
